+++ server/models/product.js
@@ -9,7 +9,7 @@ const productModel = (sequelize, DataTypes) => {
         },
         product_name: DataTypes.STRING,
         image_url: DataTypes.STRING,
-        price: DataTypes.NUMERIC,
+        price: DataTypes.INTEGER,
         },
         {
             sequelize,
